minibuh/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
aiogram==3.6.0
python-dotenv==1.0.1
SQLAlchemy==2.0.30
pydantic==2.8.2
pydantic-settings==2.3.4
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.140.0
python-docx==1.1.2
pdfplumber==0.11.4
httpx==0.27.0
aiohttp==3.9.5
